pyproject.toml
[project]
name = "kindle-capture"
version = "0.1.0"
description = "Kindle for Mac screenshot capture tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "playwright>=1.58.0",
    "pypdf>=6.6.2",
    "reportlab>=4.4.9",
    "yomitoku>=0.10.3",
]

[tool.setuptools]
py-modules = ["step1"]
